+++ src/type/type.controller.ts
@@ -8,6 +8,7 @@ import {
   Patch,
   Post,
 } from '@nestjs/common';
+import { NotFoundException } from '@nestjs/common/exceptions';
 import { CreateTypeDto, UpdateTypeDto } from './dto';
 import { TypeService } from './type.service';
 
@@ -30,11 +31,19 @@ export class TypeController {
     @Param('id', ParseIntPipe) id: number,
     @Body() payload: UpdateTypeDto,
   ) {
+    const type = this.typeService.getTypeById(id);
+    if (!type) {
+      throw new NotFoundException('Type not found');
+    }
     return this.typeService.updateType(id, payload);
   }
 
   @Delete(':id')
   deleteType(@Param('id', ParseIntPipe) id: number) {
+    const type = this.typeService.getTypeById(id);
+    if (!type) {
+      throw new NotFoundException('Type not found');
+    }
     return this.typeService.deleteType(id);
   }
 }
